# task-management/entity-apis/task-status/openapi.yaml
openapi: 3.0.0
servers:
  - description: The local dev server
    url: http://localhost:8080/taskstatus
info:
  description: "CRUD Operations for Tasks Status"
  version: "1.0.0"
  title: "Task Status API"
  contact:
    email: "choreo-help@wso2.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
tags:
  - name: "todo"
    description: "Sample todo application"
  - name: "demo"
    description: "This api is created for demo purposes"
paths:
  /taskstatus:
    get:
      summary: "returns task status for a given user id"
      operationId: "getTasksStatusByUserId"
      description: |
        Give a list of task status given the user Id. list will be sorted by last updated date.
      parameters:
        - in: "query"
          name: "userId"
          description: "User Id"
          required: true
          schema:
            type: string
      responses:
        "200":
          description: "list of tasks statuses belonging to the given user"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TaskStatus"
        "500":
          description: "internal error occurred"
    post:
      summary: "Create a Task Status"
      operationId: "createTaskStatus"
      description: "Create a task status with given details"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: "InProgress"
                userId:
                  type: string
                  example: "698d8383-28f1-4d2c-8179-e11a290e0910"
        description: "An example task status"
      responses:
        "201":
          description: "item created"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskStatus"
        "400":
          description: "invalid input, object invalid"
        "500":
          description: "internal server error occurred"
  /taskstatus/{id}:
    get:
      summary: "Retrieve a task status by id"
      operationId: "getTaskStatusById"
      description: |
        Retrieve a task status by id.
      parameters:
        - name: id
          in: path
          description: "Task Status Id"
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: "The task status with Id"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskStatus"

        "404":
          description: "Unable to find the task status"
    put:
      summary: "Update a task status by id"
      operationId: "updateTaskStatusById"
      description: |
        Update a task Status by id.
      parameters:
        - name: id
          in: path
          description: "Task Status Id"
          required: true
          schema:
            type: integer
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: "InProgress"
        description: "An example task status"
      responses:
        "200":
          description: "The task status with Id"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskStatus"

        "404":
          description: "Unable to find the task status"
    delete:
      summary: "Delete a task status by id"
      operationId: "deleteTaskStatusById"
      description: |
        Delete a task status by id.
      parameters:
        - name: id
          in: path
          description: "Task Status Id"
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: "The task status with Id"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskStatus"
        "404":
          description: "Unable to find the task status"
components:
  schemas:
    Task:
      type: object
      required:
        - id
        - title
        - taskGroupId
        - taskStatus
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
          minimum: 1
          example: 10
        title:
          type: string
          example: "Review monthly expenses"
        taskGroupId:
          type: integer
          minimum: 1
          example: 5
        taskStatus:
          type: string
          example: "Open"
        createdAt:
          type: string
          format: date-time
          example: "2022-10-12T15:56:12.171Z"
        updatedAt:
          type: string
          format: date-time
          example: "2022-10-12T15:56:12.171Z"
    TaskGroup:
      type: object
      required:
        - id
        - title
        - userId
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
          minimum: 1
          example: 10
        title:
          type: string
          example: "Urgent and important tasks"
        userId:
          type: string
          example: "ea69ddda-c44f-4ba5-a701-fda8a6c1ff42"
        createdAt:
          type: string
          format: date-time
          example: "2022-10-12T15:56:12.171Z"
        updatedAt:
          type: string
          format: date-time
          example: "2022-10-12T15:56:12.171Z"
    TaskStatus:
      type: object
      required:
        - id
        - name
        - userId
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
          minimum: 1
          example: 10
        name:
          type: string
          example: "InProgress"
        userId:
          type: string
          example: "ea69ddda-c44f-4ba5-a701-fda8a6c1ff42"
        createdAt:
          type: string
          format: date-time
          example: "2022-10-12T15:56:12.171Z"
        updatedAt:
          type: string
          format: date-time
          example: "2022-10-12T15:56:12.171Z"